fix(core): fix styles for active account tab
