+++ .changeset/cold-timers-burn.md
@@ -1,0 +1,5 @@
+---
+"@bigcommerce/catalyst-core": patch
+---
+
+fix styles for active account tab
