+++ core/app/[locale]/(default)/account/(tabs)/layout.tsx
@@ -2,16 +2,12 @@ import { useTranslations } from 'next-intl';
 import { setRequestLocale } from 'next-intl/server';
 import { PropsWithChildren } from 'react';
 
-import { Link } from '~/components/link';
 import { LocaleType } from '~/i18n/routing';
-import { cn } from '~/lib/utils';
 
-const tabList = ['addresses', 'settings'] as const;
-
-export type TabType = (typeof tabList)[number];
+import { TabNavigation } from './_components/tab-navigation';
 
 interface Props extends PropsWithChildren {
-  params: { locale: LocaleType; tab?: TabType };
+  params: { locale: LocaleType };
 }
 
 export default function AccountTabLayout({ children, params: { locale } }: Props) {
@@ -19,30 +15,10 @@ export default function AccountTabLayout({ children, params: { locale } }: Props
 
   const t = useTranslations('Account.Home');
 
-  const tabsTitles = {
-    addresses: t('addresses'),
-    settings: t('settings'),
-  };
-
   return (
     <>
       <h1 className="my-8 text-4xl font-black lg:my-8 lg:text-5xl">{t('heading')}</h1>
-      <nav aria-label={t('accountTabsLabel')}>
-        <ul className="mb-8 flex items-start overflow-x-auto">
-          {tabList.map((tab) => (
-            <li key={tab}>
-              <Link
-                className={cn('block whitespace-nowrap px-4 pb-2 font-semibold')}
-                href={`/account/${tab}`}
-                prefetch="viewport"
-                prefetchKind="full"
-              >
-                {tabsTitles[tab]}
-              </Link>
-            </li>
-          ))}
-        </ul>
-      </nav>
+      <TabNavigation />
       {children}
     </>
   );
